perdu

+++ src/staticThing.java
@@ -7,7 +7,7 @@ public class staticThing {
     private double x,y,cx1,cy1,cx2,cy2;
     private ImageView image;
 
-    public staticThing(int x, int y, String fileName,double cx1, double cy1 ,double cx2,double cy2) {
+    public staticThing(double x, double y, String fileName,double cx1, double cy1 ,double cx2,double cy2) {
         this.x=x;
         this.y=y;
         this.cx1=cx1;
